PR comments: graph utils refactoring
1. More meaningful names: `GraphBuilderSession` => `IsolatedSession`
2. Tensorflow graph utils resides in a single module
   (as opposed to a static class)
3. `GraphFunction` static methods
