+++ python/tests/graph/__init__.py
@@ -1,3 +1,4 @@
+#
 # Copyright 2017 Databricks, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,16 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-from .image.imageIO import imageSchema, imageType, readImages
-from .transformers.keras_image import KerasImageFileTransformer
-from .transformers.named_image import DeepImagePredictor, DeepImageFeaturizer
-from .transformers.tf_image import TFImageTransformer
-from .transformers.utils import imageInputPlaceholder, stripAndFreezeGraph
-
-__all__ = [
-    'imageSchema', 'imageType', 'readImages',
-    'TFImageTransformer',
-    'DeepImagePredictor', 'DeepImageFeaturizer',
-    'KerasImageFileTransformer',
-    'imageInputPlaceholder', 'stripAndFreezeGraph']
